ajout blocage des users non autorisés selon le role

+++ controllers/categoryControllers.js
@@ -2,21 +2,28 @@ import CategoryModel from '../models/categoryModel.js'
 import AppError from '../utils/AppError.js'
 
 export const addCategory = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const category = new CategoryModel(req.body)
     const isAlreadyExist = await CategoryModel.findOne({ name_cat: category.name_cat })
+
     await category.save()
     if (isAlreadyExist) {
         return next(new AppError(`Cette catégorie existe déjà.`, 400))
-    } else if (!category || !category.name_cat || !category.description_cat || !category.max_weight || !category.max_altitude || !category.fly_type || !category.key) {
-        return next(new AppError(`Erreur lors de la création de la catégorie.`, 400))
-    } else {
-        res.status(201).send({
-            message: 'Catégorie créée avec succès',
-            category: category
-        }
-        )
     }
+    if (!category || !category.name_cat || !category.description_cat || !category.max_weight || !category.max_altitude || !category.fly_type || !category.key) {
+        return next(new AppError(`Erreur lors de la création de la catégorie.`, 400))
+    }
+    res.status(201).send({
+        message: 'Catégorie créée avec succès',
+        category: category
+    }
+    )
 }
+
+
 
 export const getAllCategories = async (req, res, next) => {
     const categories = await CategoryModel.find({})
@@ -36,20 +43,30 @@ export const getCategory = async (req, res, next) => {
 }
 
 export const updateCategory = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const category = await CategoryModel.findByIdAndUpdate(req.params.idCategory, req.body)
-    await category.save()
+    
     if (!category || category === null || category === undefined || category === '') {
         return next(new AppError(`Aucune catégorie trouvée.`, 404))
-    } else {
-        res.send({
-            message: `Catégorie ${category.name_cat} modifiée avec succès.`,
-            category: category
-        })
     }
+    
+    await category.save()
+    res.send({
+        message: `Catégorie ${category.name_cat} modifiée avec succès.`,
+        category: category
+    })
 }
 
 export const deleteCategory = async (req, res, next) => {
+    if (req.user.key_r !== 1) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const category = await CategoryModel.findByIdAndDelete(req.params.idCategory)
+
     if (!category || category === null || category === undefined || category === '') {
         return next(new AppError(`Aucune catégorie trouvée.`, 404))
     } else {
