+++ controllers/droneControllers.js
@@ -1,20 +1,24 @@
 import DroneModel from '../models/droneModel.js'
 import AppError from '../utils/AppError.js'
 
-export const addDrone = async (req, res) => {
+export const addDrone = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+    
     const drone = new DroneModel(req.body)
     await drone.save()
     if (!drone || drone === null || drone === undefined || drone === '') {
         return next(new AppError(`Erreur lors de la création du drone.`, 400))
-    } else {
-        res.status(201).send({
-            message: 'Drone créé avec succès',
-            drone: drone
-        })
     }
+    res.status(201).send({
+        message: 'Drone créé avec succès',
+        drone: drone
+    })
 }
 
-export const getAllDrones = async (req, res) => {
+
+export const getAllDrones = async (req, res, next) => {
     const drones = await DroneModel.find({})
         .populate('name_cat')
     res.send(drones)
@@ -34,8 +38,6 @@ export const getDrone = async (req, res, next) => {
         })
 }
 
-
-
 export const getDronesByCategory = async (req, res, next) => {
     const drone = await DroneModel.find({ category_id: req.params.idCategory })
         .populate('category_id')
@@ -47,6 +49,10 @@ export const getDronesByCategory = async (req, res, next) => {
 }
 
 export const updateDrone = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const drone = await DroneModel.findByIdAndUpdate(req.params.idDrone, req.body)
     if (!drone || drone.length === 0 || drone === null || drone === undefined || drone === '') {
         return next(new AppError(`Aucun drone ${req.params.idDrone} trouvé.`, 404))
@@ -59,6 +65,10 @@ export const updateDrone = async (req, res, next) => {
 }
 
 export const deleteDrone = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const drone = await DroneModel.findByIdAndDelete(req.params.idDrone)
     if (!drone || drone.length === 0 || drone === null || drone === undefined || drone === '') {
         return next(new AppError(`Aucun drone ${req.params.idDrone} trouvé.`, 404))
