+++ controllers/orderControllers.js
@@ -2,7 +2,10 @@ import OrderModel from '../models/orderModel.js'
 import DroneModel from '../models/droneModel.js'
 import AppError from '../utils/AppError.js'
 
-export const addOrder = async (req, res) => {
+export const addOrder = async (req, res, next) => {
+    if (!req.user.key_r) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
     const order = new OrderModel(req.body)
     await order.save()
     res.status(201).send({
@@ -11,18 +14,25 @@ export const addOrder = async (req, res) => {
     })
 }
 
-export const getAllOrders = async (req, res) => {
+export const getAllOrders = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
     const orders = await OrderModel.find({})
     res.send(orders)
 }
 
 export const getOrderById = async (req, res, next) => {
+
+    if (!req.user.key_r) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
     const order = await OrderModel.findById(req.params.idOrder)
         .populate('user_id')
         .populate('drone_id')
         .exec((err, order) => {
             if (err) {
-                return next(new AppError(`Erreur lors de la récupération de la commande ${req.params.idOrder}`, 400))
+                return next(new AppError('Erreur lors de la récupération de la commande', 400))
             }
             if (!order || order === null || order === undefined || order === '') {
                 return next(new AppError(`Aucune commande ${req.params.idOrder} trouvée.`, 404))
@@ -35,23 +45,33 @@ export const getOrderById = async (req, res, next) => {
 }
 
 export const getOrdersByUserId = async (req, res, next) => {
-    const order = await OrderModel.find({ user_id: req.params.idUser })
-        .populate('drone_id')
-        .exec((err, order) => {
-            if (err) {
-                return next(new AppError(`Erreur lors de la récupération de la commande ${req.params.idUser}`, 400))
-            }
-            if (!order || order === null || order === undefined || order === '' || order.length === 0) {
-                return next(new AppError(`Aucune commandes trouvées de l'utilisateur ${req.params.idUser}`, 404))
-            }
-            res.status(200).send({
-                message: `Commandes trouvées pour l'utilisateur ${req.params.idUser} `,
-                order
+    if (req.user.key_r) {
+        if (req.user.key_r === 3 && req.user._id !== req.params.idUser) {
+            return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+        }
+
+        const order = await OrderModel.find({ user_id: req.params.idUser })
+            .populate('drone_id')
+            .exec((err, order) => {
+                if (err) {
+                    return next(new AppError('Erreur lors de la récupération de la commande', 400))
+                }
+                if (!order || order === null || order === undefined || order === '' || order.length === 0) {
+                    return next(new AppError(`Vous n'avez pas de commandes en cours.`, 404))
+                }
+                res.status(200).send({
+                    message: `Voici vos commandes : `,
+                    order
+                })
             })
-        })
+    }
 }
 
 export const updateOrder = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const order = await OrderModel.findByIdAndUpdate(req.params.idOrder, req.body)
     if (!order || order === null || order === undefined || order === '') {
         return next(new AppError(`Aucune commande ${req.params.idOrder} trouvée.`, 404))
@@ -61,7 +81,6 @@ export const updateOrder = async (req, res, next) => {
     if (!drone || drone === null || drone === undefined || drone === '') {
         return next(new AppError(`Aucun drone ${order.drone_id} trouvé.`, 404))
     }
-    
     await drone.save()
     res.send({
         message: `Commande ${req.params.idOrder} modifiée avec succès`,
@@ -71,8 +90,12 @@ export const updateOrder = async (req, res, next) => {
 }
 
 export const deleteOrder = async (req, res, next) => {
+    if (req.user.key_r !== 1) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const order = await OrderModel.findByIdAndDelete(req.params.idOrder)
-    if (!order) {
+    if (!order || order === null || order === undefined || order === '') {
         return next(new AppError(`Aucune commande ${req.params.idOrder} trouvée.`, 404))
     }
     res.status(200).send({
