+++ controllers/qrCodeController.js
@@ -3,6 +3,10 @@ import AppError from '../utils/AppError.js'
 import qrcode from 'qrcode'
 
 export const getQrCode = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const qrCode = await qrCodeModel.findById(req.params.idQrCode)
     if (!qrCode) {
         return next(new AppError(`Aucun QrCode ${req.params.idQrCode} trouvé.`, 404))
@@ -11,6 +15,10 @@ export const getQrCode = async (req, res, next) => {
 }
 
 export const addQrCode = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const qrCode = new qrCodeModel(req.body)
     const url = req.body.src
 
@@ -24,7 +32,7 @@ export const addQrCode = async (req, res, next) => {
             qrCode.save()
                 .then(() => {
                     res.status(201).send({
-                        message: 'QrCode created',
+                        message: 'QR Code créé.',
                         qrCode: qrCode
                     })
                 })
@@ -38,21 +46,28 @@ export const addQrCode = async (req, res, next) => {
 }
 
 
-export const getAllQrCodes = async (req, res) => {
-    const qrCodes = await qrCodeModel.find({})
-    if(!qrCodes || qrCodes.length === 0 || qrCodes === null || qrCodes === undefined || qrCodes === ''){
-        return next(new AppError(`Aucun QrCode trouvé.`, 404))
+export const getAllQrCodes = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
     }
+    const qrCodes = await qrCodeModel.find({})
+    if (!qrCodes || qrCodes.length === 0 || qrCodes === null || qrCodes === undefined || qrCodes === '')
+        return next(new AppError(`Aucun QrCode trouvé.`, 404))
+
     res.send(qrCodes)
 }
 
 export const deleteQrCode = async (req, res, next) => {
+    if (req.user.key_r > 2) {
+        return next(new AppError(`Vous n'êtes pas autorisé à effectuer cette action.`, 403))
+    }
+
     const qrCode = await qrCodeModel.findByIdAndDelete(req.params.idQrCode)
-    console.log(qrCode)
+
     if (!qrCode || qrCode.length === 0 || qrCode === null || qrCode === undefined || qrCode === '') {
         return next(new AppError(`Aucun QrCode ${req.params.idQrCode} trouvé.`, 404))
     }
     res.status(200).send({
-        message: `QR Code ${req.params.idQrCode} supprimé.`
+        message: `QR Code supprimé.`
     })
 }
